+++ tests/integration/cases/call_per_element_false/VisualBasic_call.vb
@@ -1,0 +1,8 @@
+Module Check
+    Function process(data As Object) As Object
+        Return Nothing
+    End Function
+    Sub _calls()
+        process(1)
+    End Sub
+End Module